chore: rename pending transaction footer component files
