+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/common/pending_tx_actions_footer.style.tsx
