+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/common/pending_tx_actions_footer.tsx
@@ -1,7 +1,7 @@
-// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// Copyright (c) 2024 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
 
@@ -25,7 +25,10 @@ import {
   LoadIcon,
   QueueStepButton
 } from './style'
-import { FooterButtonRow, rejectAllButtonRowPadding } from './footer.style'
+import {
+  FooterButtonRow,
+  rejectAllButtonRowPadding
+} from './pending_tx_actions_footer.style'
 
 interface Props {
   rejectButtonType?: 'reject' | 'cancel'
